+++ ChangeLog.md
@@ -766,7 +766,17 @@ Exit criteria:
 
 25. Add a cancelBuilding method to the Engine, to be activated by the back button if a building is selected. Clicking 'back' when a building is selected does not change the displayed building option buttons, but cancels the placement selection process for the currently selected building (sets the mouse context back to select and sets the selected building value to null).
 
-### 26. Iron out the kinks in building rendering and placement so that buildings don't disappear until they're entirely outside of the screen (right now they disappear as soon as one of their columns is outside the game's render area). Also, the building placement shadow should only be rendered if the mouse is over the map area, not the sidebar.
+### 26. Iron out the kinks in building rendering and placement process:
+
+- Buildings don't disappear until they're entirely outside of the screen
+- Building placement shadow should only be rendered if the mouse is over the map area, not the sidebar
+- Add an outline to selected build chips, so that the player can see which building is selected
+- Disable map scrolling when mouse context is in building placement mode
+- Clicking the building chip for the building that is already selected unselects it
+
+### 27. Clean up the on-screen 'console log' text before ending the chapter.
+
+## Chapter Nineteen: Buildings in the Frontend, Part II (Difficulty Estimate: 6 for new dynamics development)
 
 ## Chapter X: Loading Games (Forecast now says it'll be chapter 18...)